# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the api documentation of task manager project.",
        "title": "Task Manager API Documentation",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9090",
    "basePath": "/",
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "description": "UserLogin API is made for login the user in task manager application.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email of the user",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Password of the user",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User login done successfully.",
                        "schema": {
                            "$ref": "#/definitions/response.UserWithTokens"
                        }
                    },
                    "400": {
                        "description": "Bad request.",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "401": {
                        "description": "Password not matched.",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "404": {
                        "description": "User not found.",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/registration": {
            "post": {
                "description": "UserRegistration API is made for registering a new user in the task manager application.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "First name of the user",
                        "name": "firstName",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Last name of the user",
                        "name": "lastName",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bio of the user",
                        "name": "bio",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Email of the user",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Password of the user",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "privacy of the user (PUBLIC, PRIVATE)",
                        "name": "privacy",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User created successfully.",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request.",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "409": {
                        "description": "Duplicate email found.",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/reset-token": {
            "post": {
                "description": "ResetToken API is made for reset the user's access token.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Reset Access Token",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003crefresh_token\u003e",
                        "description": "Refresh Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token refresh done successfully.",
                        "schema": {
                            "$ref": "#/definitions/response.Tokens"
                        }
                    },
                    "401": {
                        "description": "Either refresh token not found or token is expired.",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.CustomError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errorhandling.CustomError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "Bad request"
                },
                "error": {
                    "type": "string",
                    "example": "Corresponding Error Message will Show Here"
                }
            }
        },
        "response.SuccessResponse": {
            "description": "Send success response to client with corresponding message and id(if any).",
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "200 OK"
                },
                "id": {
                    "type": "integer",
                    "example": 974751326021189896
                },
                "message": {
                    "type": "string",
                    "example": "Task Created Successfully."
                }
            }
        },
...
        "response.Tokens": {
            "description": "Used to send access and refresh token to response.",
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                },
                "code": {
                    "type": "string",
                    "example": "200 OK"
                },
                "refreshToken": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                }
            }
        },
        "response.User": {
            "description": "User information with id, first name, last name, bio, email, password and privacy.",
            "type": "object",
            "properties": {
                "bio": {
                    "type": "string",
                    "example": "Junior Software Engineer at ZURU TECH INDIA."
                },
                "email": {
                    "type": "string",
                    "example": "chiragmakwana@gmail.com"
                },
                "firstName": {
                    "type": "string",
                    "example": "Chirag"
                },
                "id": {
                    "type": "integer",
                    "example": 974751326021189896
                },
                "lastName": {
                    "type": "string",
                    "example": "Makwana"
                },
                "password": {
                    "type": "string",
                    "example": "Chirag123$,omitempty"
                },
                "privacy": {
                    "type": "string",
                    "example": "PUBLIC"
                }
            }
        },
        "response.UserWithTokens": {
            "description": "Send user info along with access token and refresh token to response.",
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                },
                "code": {
                    "type": "string",
                    "example": "200 OK"
                },
                "refreshToken": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                },
                "user": {
                    "$ref": "#/definitions/response.User"
                }
            }
        }
    }
}